create feedHandler
